Add files via upload
I wrote this code to aid my German learning process. This code helps me find out if the new word I just learned exists in any of the AnnenMayKantereit songs I entered. If it does, it tells me which one. It’s a very simple code, and we can't really say it’s highly functional, but as a beginner, I wrote this code, and honestly, it has been helpful to me.